requirement.txt
alembic==1.18.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
Authlib==1.6.5
azure-common==1.1.28
azure-core==1.38.0
azure-identity==1.25.1
azure-mgmt-core==1.6.0
azure-mgmt-rdbms==10.1.1
azure-mgmt-resource==24.0.0
azure-mgmt-subscription==3.1.1
babel==2.17.0
backports.zstd==1.3.0
bcrypt==5.0.0
bidict==0.23.1
blinker==1.9.0
boto3==1.42.27
botocore==1.42.27
brotli==1.2.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.9.0
fastar==0.8.0
Flask==3.1.2
flask-babel==4.0.0
Flask-Compress==1.23
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.7.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.1
httptools==0.7.1
httpx==0.28.1
idna==3.11
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
jsonformatter==0.3.4
keyring==25.7.0
ldap3==2.9.1
libgravatar==1.0.4
libpass==1.9.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
oauthlib==3.3.1
paramiko==3.5.1
passlib==1.7.4
pgadmin4==9.11
pillow==12.1.0
proto-plus==1.27.0
protobuf==6.33.4
psutil==7.1.3
psycopg==3.3.0
psycopg-binary==3.3.0
psycopg2==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.2
pyotp==2.9.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.0
python-jose==3.5.0
python-multipart==0.0.21
python-socketio==5.16.0
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rsa==4.9.1
s3transfer==0.16.0
SecretStorage==3.5.0
sentry-sdk==2.49.0
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.45
sqlmodel==0.0.31
sqlparse==0.5.5
sshtunnel==0.4.0
starlette==0.50.0
typer==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ua-parser==1.0.1
ua-parser-builtins==202601
uritemplate==4.2.0
urllib3==2.6.3
user-agents==2.2.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.5
wsproto==1.3.2
WTForms==3.2.1
